Move lettre into module
